% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANNt_versus_Buffet_Experiment.R
\name{ANNt_versus_Buffet_Experiment}
\alias{ANNt_versus_Buffet_Experiment}
\title{ANNt_versus_Buffet_Experiment
Realize the 1st time series research, Validation: Long Term Testing, defined in "Value investing or quantitative financing: portfolio decision based on a new efficient frontier concept" paper.}
\usage{
ANNt_versus_Buffet_Experiment(
  Data_Base = "GitHub",
  Serie = "First_Serie",
  Initial_Date = "2018-01-03",
  Final_Date_Training = "2021-12-30",
  Final_Date = "2022-08-04 ",
  Hidden = Hidden,
  Stepmax = Stepmax,
  Loss = "MSE",
  Learning_Rate = 0.3,
  Decay = "No",
  Early_Stopping = "No",
  Asymmetry = "Negative",
  Skew_t = "No"
)
}
\arguments{
\item{Data_Base}{import from Yahoo Finance or if "GitHub" to be specified, from GitHub/aleoli05}

\item{Serie}{"First_Serie" or 'Rebalanced_Serie" for the first or rebalanced sample}

\item{Initial_Date}{Series start Date (Must be 7 periods greater than the analyzed series)}

\item{Final_Date_Training}{Series finish training date}

\item{Final_Date}{Series end Date (If '' is the System Date)}

\item{Hidden}{Number of hidden neurons (If '' is the length series)}

\item{Stepmax}{Number of replications per asset to train the ANN}

\item{Loss}{Function: "MSE" for Mean Square Error, "MAE" for Mean Absolute Error,
"MADL" for Mean Absolute Directional Loss, and "GMADL" for Generalized Mean Absolute Directional Loss}

\item{Learning_Rate}{is the Artificial Neural Network learning rate}

\item{Decay}{L2 regularization or weight decay, add a penalty term to the loss function. "Yes" or "No.
No" is default. If  "Yes" is necessary inform the lambda or rate of regularization}

\item{Early_Stopping}{= 'No' or 'Yes'. Default is 'No'. If 'Yes' is necessary inform the value}

\item{Asymmetry}{"Negative" or "Positive". Shifts the probability of the return being greater than the proxy to the right or left, "Negative" or "Positive". Default is to the right, "Negative"}

\item{Skew_t}{Incorporate skew parameter in the probability: "Yes" or "No". Default is "No".}

\item{Parameters}{was defined in paper}
}
\description{
ANNt_versus_Buffet_Experiment
Realize the 1st time series research, Validation: Long Term Testing, defined in "Value investing or quantitative financing: portfolio decision based on a new efficient frontier concept" paper.
}
\examples{
ANNt_versus_Buffet_Experiment(Data_Base="GitHub",
Serie="First_Serie",
Initial_Date='2018-01-11',
Final_Date_Training='2021-12-30',
Final_Date='2022-08-04 ',
Hidden= Hidden, Stepmax= Stepmax,
Loss="MSE", Learning_Rate=0.3, Decay='No',
Early_Stopping = 'No', Asymmetry='Negative',
Skew_t='No')

}
\author{
Alexandre Silva de Oliveira
}
