% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assets_series_Rus.R
\name{Assets_series_Rus}
\alias{Assets_series_Rus}
\title{Assets series Rus
Import the series with the length specified of the Moscow Stock Exchange}
\usage{
Assets_series_Rus(Tickers, RM, Initial_Date, Final_Date, Periodicity)
}
\arguments{
\item{Tickers}{Name of the assets}

\item{RM}{Proxy of the Russia market - Example: IMOEX - Moscow Exchange Index}

\item{Initial_Date}{Series start Date, format ('Year-Month-Day')

Assets with values not observed in the series are excluded}

\item{Final_Date}{Series end Date ('Year-Month-Day'). If '' is the System Date}

\item{Periodicity}{should be one of “daily”, “weekly”, “monthly”, “hourly”, “1minutes”, “2minutes”, “5minutes”, “15minutes”, “30minutes”, “60minutes”, “90minutes”. (Intraday maximum 7 days)}
}
\description{
Assets series Rus
Import the series with the length specified of the Moscow Stock Exchange
}
\examples{
# Specify the assets or "Current_SP500_Tickers" for all S&P 500 assets
Tickers <-c('AFKS', 'AFLT', 'FEES', 'GMKN','GAZP')
RM <-c('^GSPC') #RM the S&P500
Initial_Date <-c('2018-01-03')
Final_Date <-c('2023-09-07')
Periodicity <- c('daily')

# Generates the Adjusted Daily Prices Series from Yahoo Finance
Assets_series_Rus (Tickers=c('AFKS', 'AFLT', 'FEES', 'GMKN','GAZP'),'IMOEX', '2018-01-03', '','daily')

}
