% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cost_transactions.R
\name{Cost_transactions}
\alias{Cost_transactions}
\title{Cost_transactions}
\usage{
Cost_transactions(Cost = 0.05)
}
\arguments{
\item{Cost}{Percentual of cost transactions consider. This number will be multiplied for 2 because we need sell and buy.}
}
\description{
Penalize the return of portfolios with the cost transactions
}
\examples{
Const_transactions(Cost=0.05)
}
