% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learning_Curve.R
\name{Learning_Curve}
\alias{Learning_Curve}
\title{Learning_Curve
 classify assets by the probability of return exceeding a RM and present the Learning Curve
Use Artificial Neural Networks (ANN) and t-distribution, the number of ANN is the number of import assets
the graphs are save in the document past}
\usage{
Learning_Curve(
  Initial_Date_Training,
  Final_Date_Training,
  Final_Date_Testing,
  N_Lags = 5,
  Hidden,
  Stepmax,
  Loss = "MSE",
  Learning_Rate = 0.3,
  Decay = "No",
  Early_Stopping = "No",
  Asymmetry = "Negative",
  Skew_t = "No",
  Prediction = "Predict",
  Bias = "No",
  Order_Only = "No",
  Convolution = "Neutral",
  Initialization = "Original",
  Activation_Function = "Tangent",
  Activation_F_Out = "Original",
  Batch_Size = ""
)
}
\arguments{
\item{Initial_Date_Training}{Series start Date (Must be 7 periods greater than the analyzed series)}

\item{Final_Date_Training}{Series finish training date}

\item{Final_Date_Testing}{Series end Date (If '' is the System Date)}

\item{N_Lags}{Number of lags used in the input layer}

\item{Hidden}{Number of hidden neurons (If '' is the length series)}

\item{Stepmax}{Number of replications per asset to train the ANN}

\item{Loss}{Function: "MSE" for Mean Square Error, "MAE" for Mean Absolute Error,
"MADL" for Mean Absolute Directional Loss, and "GMADL" for Generalized Mean Absolute Directional Loss}

\item{Learning_Rate}{is the Artificial Neural Network learning rate}

\item{Decay}{L2 regularization or weight decay, add a penalty term to the loss function. "Yes" or "No.
No" is default. If  "Yes" is necessary inform the lambda or rate of regularization}

\item{Early_Stopping}{= 'No' or 'Yes'. Default is 'No'. If 'Yes' is necessary inform the value}

\item{Asymmetry}{"Negative" or "Positive". Shifts the probability of the return being greater than the proxy to the right or left, "Negative" or "Positive". Default is to the right, "Negative"}

\item{Skew_t}{Incorporate skew parameter in the probability: "Yes" or "No". Default is "No".}

\item{Prediction}{Model of prediction: "Predict" or "Forecast". "Predict" is the standard.}

\item{Bias}{include Bias, Yes or No, with auto learning}

\item{Order_Only}{disability the ANN and only order the historic probability to outperformed the benchmark}

\item{Convolution}{addresses the bearish/bullish tendency or inverse tendency in the neural input (Trend, Neutral, Reverse)}

\item{Initialization}{define the parameters of weights initialization:
1) Original: Define the first configuration. Uniform distribution and weights from 0 to 1;
2) Xavier_N: Define the Normal distribution with variance of Xavier;
3) Xavier_U: Define the Uniform distribution with limit's Xavier;
4) Xavier_H: Define HeLu limit's Xavier;
5) Xavier_Leaky: Define Leaky HeLU limit's Xavier
6) Normal_D: Define the Normal distribution with mean=0 and Variance=1;
7) Uniform_D: Define the Uniform distribution with minimum=-0.5 and maximum=0.5 for stability.}

\item{Activation_Function}{Define the Activation Function
1) Tangent: Define the Tangent Function Activation
2) Sigmoid: Define the Sigmoid Function Activation
3) Leaky_ReLU: Define the Leaky ReLU Function Activation}

\item{Activation_F_Out}{Define the Activation Function:
1) Original: Define the Tangent function activation and Sigmoid Derivate Backpropagation
2) Linear: Define the linear function activation in the output layer
3) Activation: Define the Activation function in the output layer specified in the hidden layers}

\item{Batch_Size}{Define the length of Batch. Standard is full batch size (Batch gradient descent)}
}
\description{
Learning_Curve
 classify assets by the probability of return exceeding a RM and present the Learning Curve
Use Artificial Neural Networks (ANN) and t-distribution, the number of ANN is the number of import assets
the graphs are save in the document past
}
\examples{
Learning_Curve(
Initial_Date_Training = c('2018-01-11'),
Final_Date_Training = c('2022-12-29'),
Final_Date_Testing = c(''),
N_Lags=5,
Hidden = 7,
Stepmax = 300,
Loss = "MSE",
Learning_Rate = 0.001,
Decay=c('Yes', 0.001),
Early_Stopping = c('Yes', 0.001),
Asymmetry='Negative',
Skew_t='Yes',
Bias="No",
Order_Only='No',
Convolution = 'Neutral'
)
# Estimated processing time 2 hours.

}
\author{
Alexandre Silva de Oliveira
}
