% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSTM_RNN.R
\name{LSTM_RNN}
\alias{LSTM_RNN}
\title{LSTM_RNN
Command that calculate the Long Short Term Recurent Neural Network}
\usage{
LSTM_RNN(
  Tickers = "AAPL",
  Lookback = 8,
  Initial_Date_Training = c("2018-01-03"),
  Final_Date_Training = c("2022-12-29"),
  Final_Date_Testing = c(""),
  Input_dim = 1,
  Hidden_dim = 32,
  Num_layers = 2,
  Output_dim = 1,
  Num_epochs = 100,
  Import = "Yes",
  Metric = "Return",
  Plot = "Yes"
)
}
\arguments{
\item{Tickers}{data frame with return time series}

\item{Lookback}{Sequence length to Long Short Term}

\item{Final_Date_Training}{Series end Date (If '' is the System Date)}

\item{Input_dim}{Input the dimension of the RNN}

\item{Hidden_dim}{Input the dimension of the hidden layer}

\item{Num_layers}{Number of layers}

\item{Output_dim}{Output dimension of the RNN}

\item{Num_epochs}{Number of the epochs}

\item{Import}{Import assets series}

\item{Metric}{is the "Return" of the asset or the "Excess_Return" of the asset in relation to the benchmark}

\item{Plot}{Develop the network learning graph}

\item{Inicial_Date}{Series start Date (Must be 7 periods greater than the analyzed series)}

\item{Date_Training}{Series finish training date}
}
\description{
LSTM_RNN
Command that calculate the Long Short Term Recurent Neural Network
}
\examples{
Tickers <-c('AAPL')
Lookback <- 8
Initial_Date_Training <- c('2018-01-03')
Final_Date_Training <- c('2022-12-29')
Final_Date_Testing <- c('')
Input_dim <- 1
Hidden_dim <- 32
Num_layers <- 2
Output_dim <- 1
Num_epochs <- 100
Import <- 'Yes'
Metric <- 'Return'
LSTM_RNN(Tickers, Lookback, Initial_Date_Training=c('2018-01-03'),Final_Date_Training=c('2022-12-29'),Final_Date_Testing=c(''),Input_dim = 1,Hidden_dim = 32,Num_layers = 2,Output_dim = 1,Num_epochs = 100,Import='Yes', Metric='Return')
}
