% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSTMt_order.R
\name{LSTMt_order}
\alias{LSTMt_order}
\title{classify assets by the probability of return exceeding a RM with LSTM
Use Long Short Term RNN and t-distribution, the number of input neurons is the number of import assets}
\usage{
LSTMt_order(
  Initial_Date_Training,
  Final_Date_Training,
  Final_Date_Testing,
  Hidden,
  Stepmax,
  Asymmetry = "Negative",
  View_Metrics = TRUE,
  Verbose = 1
)
}
\arguments{
\item{Final_Date_Training}{Series end Date (If '' is the System Date)}

\item{Hidden}{Number of hidden neurons (If '' is the length series)}

\item{Stepmax}{Number of replications per asset to train the ANN}

\item{Asymmetry}{"Negative" or "Positive". Shifts the probability of the return being greater than the proxy to the right or left, "Negative" or "Positive". Default is to the right, "Negative"}

\item{View_Metrics}{"True" or "False" for view realtime plot of training metrics}

\item{Verbose}{Verbosity mode (0 = silent, 1 = progress bar, 2 = one line per epoch)}

\item{Inicial_Date}{Series start Date (Must be 7 periods greater than the analyzed series)}

\item{Date_Training}{Series finish training date}
}
\description{
classify assets by the probability of return exceeding a RM with LSTM
Use Long Short Term RNN and t-distribution, the number of input neurons is the number of import assets
}
\examples{
Initial_Date_Training <-c('2018-01-11')
Final_Date_Training <- c('2022-12-29')
Final_Date_Testing <-c('')
Hidden <- 5
Stepmax <- 2500
LSTMt_order ('2018-01-11','2022-12-29','', 5, 2500)
# Estimated processing time 30 minutes per asset

}
\author{
Alexandre Silva de Oliveira
}
