% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_cumulative_returns.R
\name{Plot_Cumulative_Returns}
\alias{Plot_Cumulative_Returns}
\title{Plot_Cumulative_Returns
Present Cumulative Returns Graph}
\usage{
Plot_Cumulative_Returns(Until_Date, Legend_position = "topleft")
}
\arguments{
\item{Until_Date}{if '' is the sys.Date()}

\item{Legend_position:}{topleft;
topright;
bottomleft;
bottomrigh;
bottom;
top;
left;
right.}
}
\description{
Plot_Cumulative_Returns
Present Cumulative Returns Graph
}
\examples{
Until_Date <- '2023-09-15'
Plot_Cumulative_Returns('')
}
