% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_cumulative_returns_Horizon.R
\name{Plot_Cumulative_Returns_Horizon}
\alias{Plot_Cumulative_Returns_Horizon}
\title{Plot_Cumulative_Returns_Horizon
Present Cumulative Returns Graph for all investment horizon}
\usage{
Plot_Cumulative_Returns_Horizon(
  From_Date,
  Until_Date,
  Legend_position = "topleft",
  ANNt_Prob = "Yes"
)
}
\arguments{
\item{From_Date}{if '' is the Initial_Date}

\item{Until_Date}{if '' is the sys.Date()}

\item{ANNt_Prob}{Estimate the ANNt_Prob portfolios. "Yes" or "No". Default is "No".}

\item{Legend_position:}{topleft;
topright;
bottomleft;
bottomrigh;
bottom;
top;
left;
right.}
}
\description{
Plot_Cumulative_Returns_Horizon
Present Cumulative Returns Graph for all investment horizon
}
\examples{
Until_Date <- '2023-09-15'
Plot_Cumulative_Returns_Horizon('')
}
