% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_efficient_frontier.R
\name{Plot_efficient_frontier}
\alias{Plot_efficient_frontier}
\title{Plot_efficient_frontier
Generate the efficient frontier graph}
\usage{
Plot_efficient_frontier(ANNt_Prob = "No")
}
\arguments{
\item{ANNt_Prob}{Estimate the ANNt_Prob portfolios. "Yes" or "No". Default is "No".}

\item{No}{require parameters}
}
\description{
Plot_efficient_frontier
Generate the efficient frontier graph
}
\examples{
Plot_efficient_frontier()
}
