% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_Test_Diff_Mean_Portfolios.R
\name{t_Test_Diff_Mean_Portfolios}
\alias{t_Test_Diff_Mean_Portfolios}
\title{t_Test_Diff_Mean_Portfolios}
\usage{
t_Test_Diff_Mean_Portfolios(
  Ratio = "Annualized_Returns",
  Ratio2 = "",
  Model = 1
)
}
\arguments{
\item{Ratio}{inform the metrics to analysis:
Rm - Average_Return - Means of Returns from each portfolio;
Annualized_Returns - This is the standard ratio;
RCum - Cumulative_Returns;
Annualized_Volatility - Annualized Volatility;
Var - Variance with 95% confidence;
CVar - Conditional Variance 95% confidence;
Sharpe - Sharpe Ratio;
Alpha - Jensen's Alpha;
Beta - Covariance between portfolio and market proxy over market proxy Variance;
Sortino - Sortino Ratio;
Treynor - Treynor Ratio;
Select_Arch_RCum - Select ANN's architecture.}

\item{Ratio2}{Standard is NULL. If informed, can assume the metrics to analysis.}

\item{Model}{1 or 2 with or witout significance}
}
\description{
Analyse the difference of means portfolio´s returns. Return p-value.
}
\examples{
t_Test_Diff_Mean_Portfolios(Ratio="Treynor", Ratio2="Var", Model=1)
}
