% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tickers_SP500.R
\name{Tickers_SP500}
\alias{Tickers_SP500}
\title{Tickers_SP500}
\usage{
Tickers_SP500(Tickers = "Current_SP500_Tickers", Exclude_ticket = "")
}
\arguments{
\item{Tickers}{'Current_SP500'}

\item{Exclude_ticket}{Exclude specific tickers}
}
\description{
Return Current Tickers SP500
}
\examples{
# example code

Tickers_SP500(Tickers='Current_SP500_Tickers')
}
