% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANNt_Oliveira_Ceretta_S.R
\name{ANNt_Oliveira_Ceretta_S}
\alias{ANNt_Oliveira_Ceretta_S}
\title{ANNt_Oliveira_Ceretta_Superior
Create a portfolio to outperform the Sharpe portfolio of reported assets.}
\usage{
ANNt_Oliveira_Ceretta_S(
  Tickers,
  RM,
  Rf,
  Initial_Date,
  Final_Date_Training,
  Final_Date,
  Periodicity,
  N_Lags = 5,
  Hidden,
  Stepmax,
  Loss = "MSE",
  Learning_Rate = 0.3,
  Decay = "No",
  Early_Stopping = "No",
  Asymmetry = "Negative",
  Type_ANNt = "T4",
  N_Assets,
  Base = "yahoo",
  Import = "Yes",
  Exclude_ticket = "",
  Type_ANN = "ANNt",
  Order = "Yes",
  Skew_t = "No",
  Bias = "No",
  Order_Only = "No",
  Convolution = "Neutral",
  Initialization = "Original",
  Activation_Function = "Tangent",
  Activation_F_Out = "Original",
  Batch_Size = "",
  ANNt_Prob = ANNt_Prob,
  Delay = "No"
)
}
\arguments{
\item{Tickers}{Name of the assets or "Current_SP500_Tickers" for all S&P 500 assets}

\item{RM}{Proxy of the market}

\item{Rf}{Risk free rate}

\item{Initial_Date}{Series start Date, format ('Year-Month-Day'). Assets with values not observed in the series are excluded}

\item{Final_Date}{End date of the treatment series}

\item{Periodicity}{should be one of “daily”, “weekly”, “monthly”}

\item{N_Lags}{Number of lags used in the input layer}

\item{Hidden}{Number of hidden neurons (If ” is the length series). For a good performance use '' to form a square input x hidden matrix of neurons}

\item{Stepmax}{Number of replications per asset to train the ANN. For a good performance, use 7500}

\item{Loss}{Function: "MSE" for Mean Square Error, "MAE" for Mean Absolute Error,
"MADL" for Mean Absolute Directional Loss, and "GMADL" for Generalized Mean Absolute Directional Loss}

\item{Learning_Rate}{is the Artificial Neural Network learning rate}

\item{Decay}{L2 regularization or weight decay, add a penalty term to the loss function. "Yes" or "No.
No" is default. If  "Yes" is necessary inform the lambda or rate of regularization}

\item{Early_Stopping}{= 'No' or 'Yes'. Default is 'No'. If 'Yes' is necessary inform the value}

\item{Asymmetry}{"Negative" or "Positive". Shifts the probability of the return being greater than the proxy to the right or left, "Negative" or "Positive". Default is to the right, "Negative"}

\item{Type_ANNt}{Select type ANNt: "T1"= NNet_Signal_Traning; "T2"= NNet_t_Training; "T3"= MC_Signal_Training; "T4"= MC_t_Training; "T5"= NNet_Signal_Test; "T6"= NNet_t_Test; "T7"= MC_Signal_Test; "T8"= Type_ANNt: MC_t_Test}

\item{N_Assets}{Limit of asset numbers in the portfolio}

\item{Base}{Database to use: "yahoo" or "Rus"}

\item{Import}{Import dates from external data base after first import. "Yes"
or "No". "Yes" is the standard.}

\item{Exclude_ticket}{Deletes any ticket from the ticket list that you want to remove for some reason}

\item{Type_ANN}{Select the network type: 'ANNt', 'LSTMt' in RNN from ANNt, or 'SKEWt' to raw excess return data}

\item{Order}{If "Yes" processes the asset selection, if "No" uses the already processed assets available in the database}

\item{Skew_t}{Incorporate skew parameter in the probability: "Yes" or "No". Default is "No".}

\item{Bias}{include Bias, Yes or No, with auto learning}

\item{Order_Only}{disability the ANN and only order the historic probability to outperformed the benchmark}

\item{Convolution}{addresses the bearish/bullish tendency or inverse tendency in the neural input (Trend, Neutral, Reverse)}

\item{Initialization}{define the parameters of weights initialization:
1) Original: Define the first configuration. Uniform distribution and weights from 0 to 1;
2) Xavier_N: Define the Normal distribution with variance of Xavier;
3) Xavier_U: Define the Uniform distribution with limit's Xavier;
4) Xavier_H: Define HeLu limit's Xavier;
5) Xavier_Leaky: Define Leaky HeLU limit's Xavier
6) Normal_D: Define the Normal distribution with mean=0 and Variance=1;
7) Uniform_D: Define the Uniform distribution with minimum=-0.5 and maximum=0.5 for stability.}

\item{Activation_Function}{Define the Activation Function
1) Tangent: Define the Tangent Function Activation
2) Sigmoid: Define the Sigmoid Function Activation
3) Leaky_ReLU: Define the Leaky ReLU Function Activation}

\item{Activation_F_Out}{Define the Activation Function:
1) Original: Define the Tangent function activation and Sigmoid Derivate Backpropagation
2) Linear: Define the linear function activation in the output layer
3) Activation: Define the Activation function in the output layer specified in the hidden layers}

\item{Batch_Size}{Define the length of Batch. Standard is full batch size (Batch gradient descent)}

\item{ANNt_Prob}{generate the portfolios with ANNt probability only. Default is "No". Alternative inform: "Yes, Lambda, Num_Assets, nd nPoints}

\item{Delay}{in portfolio formation or rebalancing. Considers the number of days lag between portfolio creation and investment execution.The default is Delay="No", alternative example: Delay=c('Yes', 5)}

\item{Initial_Date_Training}{Training series start Date}
}
\description{
Command that realize all operations of the package and save all in a specific past into user past
}
\examples{
Tickers <-c('AAPL','XOM','TSLA','KO', 'F')
RM <-c('^GSPC') #RM the S&P500
Rf <- 0
Initial_Date <-c('2018-01-03')
Final_Date_Training <- c('2022-12-29')
Final_Date <-c('')
Periodicity <- c('daily')
N_Lags <- 5
Hidden <- 5
Stepmax <- 7500
Type_ANNt <- 'T8'
N_Assets <- 3
ANNt_Oliveira_Ceretta_S(c('AAPL','XOM','TSLA','KO', 'F'), '^GSPC', 0, '2018-01-03', '2022-12-29', '', 'daily',5,7500,'T8',3)
}
\author{
Alexandre Silva de Oliveira
}
