% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANNt_order.R
\name{ANNt_order}
\alias{ANNt_order}
\title{ANNt_order classify assets by the probability of return exceeding a RM
Use Artificial Neural Networks (ANN) and t-distribution, the number of ANN is the number of import assets}
\usage{
ANNt_order(
  Initial_Date_Training,
  Final_Date_Training,
  Final_Date_Testing,
  Hidden,
  Stepmax,
  Loss = "MSE",
  Learning_Rate = 0.3,
  Decay = "No",
  Early_Stopping = "No",
  Asymmetry = "Negative",
  Skew_t = "No"
)
}
\arguments{
\item{Initial_Date_Training}{Series start Date (Must be 7 periods greater than the analyzed series)}

\item{Final_Date_Training}{Series finish training date}

\item{Final_Date_Testing}{Series end Date (If '' is the System Date)}

\item{Hidden}{Number of hidden neurons (If '' is the length series)}

\item{Stepmax}{Number of replications per asset to train the ANN}

\item{Loss}{Function: "MSE" for Mean Square Error, "MAE" for Mean Absolute Error,
"MADL" for Mean Absolute Directional Loss, and "GMADL" for Generalized Mean Absolute Directional Loss}

\item{Learning_Rate}{is the Artificial Neural Network learning rate}

\item{Decay}{L2 regularization or weight decay, add a penalty term to the loss function. "Yes" or "No.
No" is default. If  "Yes" is necessary inform the lambda or rate of regularization}

\item{Early_Stopping}{= 'No' or 'Yes'. Default is 'No'. If 'Yes' is necessary inform the value}

\item{Asymmetry}{"Negative" or "Positive". Shifts the probability of the return being greater than the proxy to the right or left, "Negative" or "Positive". Default is to the right, "Negative"}

\item{Skew_t}{Incorporate skew parameter in the probability: "Yes" or "No". Default is "No".}
}
\description{
ANNt_order classify assets by the probability of return exceeding a RM
Use Artificial Neural Networks (ANN) and t-distribution, the number of ANN is the number of import assets
}
\examples{
ANNt_order (
Initial_Date_Training = c('2018-01-11'),
Final_Date_Training = c('2022-12-29'),
Final_Date_Testing = c(''),
Hidden = 7,
Stepmax = 300,
Loss = "GMADL",
Learning_Rate = 0.3,
Decay=c('Yes',0.1),
Early_Stopping = c('Yes', 0.001),
Asymmetry='Negative',
Skew_t='Yes')
# Estimated processing time 2 hours.

}
\author{
Alexandre Silva de Oliveira
}
