% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANNt_versus_Buffet_Article.R
\name{ANNt_versus_Buffet_Article}
\alias{ANNt_versus_Buffet_Article}
\title{ANNt_versus_Buffet_Article
Realize the 1st time series research, Validation: Long Term Testing, defined in "Value investing or quantitative financing: portfolio decision based on a new efficient frontier concept" paper.}
\usage{
ANNt_versus_Buffet_Article()
}
\arguments{
\item{Parameters}{was defined in paper}
}
\description{
ANNt_versus_Buffet_Article
Realize the 1st time series research, Validation: Long Term Testing, defined in "Value investing or quantitative financing: portfolio decision based on a new efficient frontier concept" paper.
}
\examples{
ANNt_versus_Buffet_Article()

}
