% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_cumulative_returns_Horizon.R
\name{Plot_Cumulative_Returns_Horizon}
\alias{Plot_Cumulative_Returns_Horizon}
\title{Plot_Cumulative_Returns_Horizon
Present Cumulative Returns Graph for all investment horizon}
\usage{
Plot_Cumulative_Returns_Horizon(Until_Date)
}
\arguments{
\item{Until_Date}{if '' is the sys.Date()}
}
\description{
Plot_Cumulative_Returns_Horizon
Present Cumulative Returns Graph for all investment horizon
}
\examples{
Until_Date <- '2023-09-15'
Plot_Cumulative_Returns_Horizon('')
}
