% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_NEF_New.R
\name{Plot_NEF_New}
\alias{Plot_NEF_New}
\title{Plot_NEF_New}
\usage{
Plot_NEF_New(
  N_Assets,
  Initial_Date_Testing,
  Final_Date_Testing,
  Rf,
  type_ANNt,
  Lambda = 0.5,
  nPoints = 200
)
}
\arguments{
\item{N_Assets}{limit of asset numbers in the portfolio}

\item{Initial_Date_Testing}{Initial Date of Test Series}

\item{Final_Date_Testing}{Final Date Test Series, if '' is the system date}

\item{Rf}{Risk free rate}

\item{type_ANNt}{Select type ANNt:
"T1"= NNet_Signal_Traning;
"T2"= NNet_t_Training;
"T3"= MC_Signal_Training;
"T4"= MC_t_Training;
"T5"= NNet_Signal_Test;
"T6"= NNet_t_Test;
"T7"= MC_Signal_Test;
"T8"= Type_ANNt: MC_t_Test}

\item{Lambda}{Risk Aversion parameter. Default is 0.5}

\item{nPoints}{number of simulation. Default is 200}
}
\description{
Plot_NEF_New
}
\examples{
N_Assets <- 3
Initial_Date_Testing <- c('2023-01-03')
Final_Date_Testing <- c('')
Rf <- 0
type_ANNt <- 'T8'
# Generate assets portfolio (maximum N assets specified)
Plot_NEF_New(3,'2023-01-03','',0,'T8')

}
\author{
Alexandre Silva de Oliveira
}
