% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Ratio_Horizon.R
\name{Plot_Ratio_Horizon}
\alias{Plot_Ratio_Horizon}
\title{Plot_Ratio_Horizon}
\usage{
Plot_Ratio_Horizon(Ratio = "Annualized_Returns")
}
\arguments{
\item{Ratio:}{Rm - Average_Return - Means of Returns from each portfolio;
Annualized_Returns - This is the standard ratio;
Cm - Cumulative_Returns;
Volatility - Annualized Volatility;
Var - Variance with 95% confidence;
CVar - Conditional Variance 95% confidence;
Sharpe - Sharpe Ratio;
Alpha - Jensen's Alpha;
Beta - Covariance between portfolio and market proxy over market proxy Variance;
Sortino - Sortino Ratio;
Treynor - Treynor Ratio.}
}
\description{
Analyzes ratios chart of portfolios with different investment horizons.
You need to run the Investment_Horizon command first.
}
\examples{
Plot_Ratio_Horizon(Ratio="Annualized_Returns")
}
