% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio_backtesting_Inv.R
\name{Portfolio_backtesting_Inv}
\alias{Portfolio_backtesting_Inv}
\title{Portfolio_backtesting_Inv
Calculate the mean portfolio backtest to the investment perod.}
\usage{
Portfolio_backtesting_Inv(Date_Initial_Backtesting, Date_Final_Backtesting)
}
\arguments{
\item{Date_Initial_Backtesting}{Date initial of the backtest}

\item{Date_Final_Backtesting}{Date final of the backtest}
}
\description{
Portfolio_backtesting_Inv
Calculate the mean portfolio backtest to the investment perod.
}
\examples{
Date_Initial_Backtesting =''
Date_Final_Backtesting =''
Portfolio_backtesting_Inv('', '')

}
