% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_Architecture.R
\name{Select_Architecture}
\alias{Select_Architecture}
\title{Select_Architecture}
\usage{
Select_Architecture(
  Tickers = c("Current_SP500_Tickers"),
  RM = c("^GSPC"),
  Rf = 0,
  Initial_Date = c("2018-01-03"),
  Final_Date_Training = c("2019-12-31"),
  Final_Date = "",
  Frequency = 2,
  Periodicity = c("daily"),
  Hidden = c("", 5, 5, 5, 5),
  Stepmax = c(2000, 300, 300, 300, 300),
  Loss = c("MSE", "MAE", "MADL", "GMADL", "GMADL"),
  Learning_Rate = c(0.3, 0.01, 0.3, 0.3, 0.01),
  Decay = c("No", c("Yes", 0.05), c("Yes", 0.5), c("Yes", 0.5), c("Yes", 0.5)),
  Early_Stopping = c("No", "Yes", "Yes", "Yes", "Yes"),
  Asymmetry = c("Positive", "Positive", "Positive", "Positive", "Positive"),
  Type_ANNt = c("T8", "T8", "T8", "T8", "T8"),
  N_Assets = c(20, 20, 20, 20, 20),
  Base = "yahoo",
  Fun = "Original",
  Specific_Dates = Specific_Dates,
  Import = "No",
  Exclude_ticket = "",
  Type_ANN = c("ANNt", "ANNt", "ANNt", "ANNt", "LSTMt"),
  Order = "Yes",
  Continue_from = "1900-01-01",
  Download = "Yes",
  Skew_t = c("No", "Yes", "Yes", "Yes", "Yes"),
  Initial_Arch = 1,
  Bias = c("No", "No", "No", "No", "No"),
  Return_Cumulative = "Total",
  Order_Only = "No"
)
}
\arguments{
\item{Tickers}{Name of the assets or "Current_SP500_Tickers" for all S&P 500 assets.}

\item{RM}{Proxy of the market.}

\item{Rf}{Risk free rate.}

\item{Initial_Date}{Series start Date, format ('Year-Month-Day'). Assets with values not observed in the series are excluded.}

\item{Final_Date}{End date of the treatment series.}

\item{Frequency}{How many times simulate the investiment horizon.}

\item{Periodicity}{should be one of “daily”, “weekly”, “monthly”.}

\item{Hidden}{Number of hidden neurons (If ” is the length series). For a good performance use '' to form a square input x hidden matrix of neurons.}

\item{Stepmax}{Number of replications per asset to train the ANN. For a good performance, use 7500.}

\item{Loss}{Function: "MSE" for Mean Square Error, "MAE" for Mean Absolute Error,
"MADL" for Mean Absolute Directional Loss, and "GMADL" for Generalized Mean Absolute Directional Loss}

\item{Learning_Rate}{is the Artificial Neural Network learning rate}

\item{Decay}{L2 regularization or weight decay, add a penalty term to the loss function. "Yes" or "No.
No" is default. If  "Yes" is necessary inform the lambda or rate of regularization}

\item{Early_Stopping}{= 'No' or 'Yes'. Default is 'No'. If 'Yes' is necessary inform the value}

\item{Asymmetry}{"Negative" or "Positive". Shifts the probability of the return being greater than the proxy to the right or left, "Negative" or "Positive". Default is to the right, "Negative"}

\item{Type_ANNt}{Select type ANNt: "T1"= NNet_Signal_Traning; "T2"= NNet_t_Training; "T3"= MC_Signal_Training; "T4"= MC_t_Training; "T5"= NNet_Signal_Test; "T6"= NNet_t_Test; "T7"= MC_Signal_Test; "T8"= Type_ANNt: MC_t_Test.}

\item{N_Assets}{Limit of asset numbers in the portfolio.}

\item{Base}{Database to use: "yahoo" or "Rus".}

\item{Fun}{Which technique to apply to generate portfolios:
'S_Out' uses the ANNt_Oliveira_Ceretta_S_Out function, this is the standard;
'Out' uses the ANNt_Oliveira_Cereta_Out function;
'S' uses the ANNt_Oliveira_Cereta_S function;
'Original' uses the ANNt_Oliveira_Ceretta function.}

\item{Import}{Import dates from external data base after first import in each re balance. "Yes"
or "No". "No" is the standard.}

\item{Exclude_ticket}{Deletes any ticket from the ticket list that you want to remove for some reason}

\item{Type_ANN}{Select the network type: 'ANNt', 'LSTMt' in RNN from ANNt, or 'SKEWt' to raw excess return data}

\item{Order}{If "Yes" processes the asset selection, if "No" uses the already processed assets available in the database}

\item{Continue_from}{Determine if continue from a Specific_Date in the data}

\item{Download}{Download asset prices for external data set or local data set. "Yes" or "No". "Yes" is the standard.}

\item{Skew_t}{Incorporate skew parameter in the probability: "Yes" or "No". Default is "No".}

\item{Initial_Arch}{Define the Architecture that will initiated the analysis: 1, 2, 3,... Standard is 1.}

\item{Bias}{include Bias, Yes or No, with auto learning}

\item{Return_Cumulative}{'Total' for all period off investment or "Rebalanced" if estimated only within the rebalancing.}

\item{Order_Only}{disability the ANN and only order the historic probability to outperformed the benchmark}

\item{Initial_Date_Training}{Training series start Date.}

\item{Specific_Date}{Specific dates for the end of training. Used to define
the investment horizon of portfolios from specific dates.}
}
\description{
Analyzes the return of portfolios with different model architecture of ANN
}
\examples{
# Specify the assets or "Current_SP500_Tickers" for all S&P 500 assets
####### Example 1 #######
Specific_Dates=c('2025-09-30','2025-06-30',
'2025-03-31','2024-12-31',
'2024-09-30','2024-06-30',
'2024-03-31','2023-12-31',
'2023-09-30','2023-06-30',
'2023-03-31','2022-12-31'
)
#######
Select_Architecture(
Tickers =c('Current_SP500_Tickers'),
RM =c('^GSPC'),
Rf = 0,
Initial_Date =c('2018-01-03'),
Final_Date_Training =c('2019-12-31'),
Final_Date ='',
Frequency = 2,
Periodicity = c('daily'),
Hidden = c('',5, 5, 5, 5),
Stepmax = c(2000, 300, 300, 300, 32),
Loss=c("MSE","MAE", "MADL", "GMADL", 'GMADL'),
Learning_Rate=c(0.3, 0.3, 0.3, 0.3, 0.3),
Decay=c('No', c('Yes',0.1), c('Yes',0.1), c('Yes',0.1), c('Yes',0.1)),
Early_Stopping = c('No','No', 'No', 'No', c('Yes', 0.001)),
Asymmetry=c('Positive','Positive','Positive','Positive','Positive'),
Type_ANNt = c('T8','T8', 'T8', 'T8', 'T8'),
N_Assets = c(20,20,20,20,20),
Base = 'yahoo',
Fun = 'Original',
Specific_Dates = Specific_Dates,
Import = 'No',
Type_ANN = c('ANNt','ANNt', 'ANNt', 'ANNt', 'LSTMt'),
Order = 'Yes',
Download = 'Yes',
Skew_t=c('No','Yes', 'Yes', 'Yes', 'Yes'),
Initial_Arch=1,
Bias=c('No','No','No','No','No'),
Return_Cumulative ='Total',
Order_Only='No'
)
}
