% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Turnovers.R
\name{Turnovers}
\alias{Turnovers}
\title{Turnovers
It measures portfolio turnover}
\usage{
Turnovers(Portfolio)
}
\arguments{
\item{Portfolios:}{Specify the portfolios or type 'Alls' for all.
Portfolios:'MF_EQ', 'MF_MKW','MKW','ANNt_EQ','ANNt_MKW','Sharpe','MF_Sharpe','ANNt_Sharpe',}
}
\description{
Turnovers
It measures portfolio turnover
}
\examples{
Turnovers('Alls')
}
