% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights_Investment_Horizon.R
\name{Weights_Investment_Horizon}
\alias{Weights_Investment_Horizon}
\title{Weights_Investment_Horizon}
\usage{
Weights_Investment_Horizon(Portfolio = "Sharpe")
}
\arguments{
\item{Portfolio:}{MF_EQ; MF_MKW; MKW; ANNt_EQ; ANNt_MKW; Sharpe; MF_Sharpe or ANNt_Sharpe}
}
\description{
Return with weights of the portfolio selected. It is necessary to run the Investiment_Horizon command before.
}
\examples{
Weights_Investment_Horizon(Portfolio='Sharpe')
}
